Update: Authorization Works Great..., +Postman collection, +TODO
Refactor for name insteed of hashes, and make hashes work

+++ JWT/src/main/java/com/example/Java4/controller/ListItemController.java
@@ -2,16 +2,25 @@ package com.example.Java4.controller;
 
 import com.example.Java4.entity.ListItem;
 import com.example.Java4.service.ListItemService;
+import com.example.Java4.service.ListService;
+import com.example.Java4.service.TokenService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
-
+import com.example.Java4.utils.Functions;
 import java.util.List;
 
 @RequestMapping("/item")
 @RestController
+
 public class ListItemController {
     @Autowired
     ListItemService listItemService;
+    @Autowired
+    ListService listService;
 
     @PostMapping("/") // localhost:8080/item
     public ListItem addListItem(@RequestBody ListItem listItem) {
@@ -19,27 +28,87 @@ public class ListItemController {
     }
 
     @PutMapping("/{itemId}/list/{listId}")  // localhost:8080/item/1/list/1
-    public ListItem assignToList(@PathVariable int itemId, @PathVariable int listId) {
-        return listItemService.assignItemToList(itemId, listId);
+    public ResponseEntity<ListItem> assignToList(@PathVariable int itemId, @PathVariable int listId, Authentication authentication) {
+        // Get the token
+        JwtAuthenticationToken jwtAuthenticationToken = (JwtAuthenticationToken) authentication;
+        String jwtToken = Functions.extractJwtToken(jwtAuthenticationToken);
+        // Get the userId
+        int userId = TokenService.getUserIdFromToken(jwtToken);
+        // Authorize
+        if (hasUserAccessToList(userId, itemId)) {
+            System.out.println("Item does belong to a user: ");
+            com.example.Java4.entity.ListItem item = listItemService.assignItemToList(itemId, listId);
+            return ResponseEntity.ok(item);
+        }
+        System.out.println("Item does NOT belong to a user: ");
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
+    // ADMIN
     @GetMapping("/s")  // localhost:8080/items
     public List<ListItem> getListItems() {
         return listItemService.getListItems();
     }
 
+    // USER
     @DeleteMapping("/{id}")  // localhost:8080/item/1
-    public String deleteItemById(@PathVariable int id) {
-        return listItemService.deleteItemById(id);
+    public ResponseEntity<ListItem> deleteItemById(@PathVariable int id, Authentication authentication) {
+        // Get the token
+        JwtAuthenticationToken jwtAuthenticationToken = (JwtAuthenticationToken) authentication;
+        String jwtToken = Functions.extractJwtToken(jwtAuthenticationToken);
+        // Get the userId
+        int userId = TokenService.getUserIdFromToken(jwtToken);
+        // Authorize
+        if (hasUserAccessToList(userId, id)) {
+            System.out.println("Item does belong to a user: ");
+            listItemService.deleteItemById(id);
+            return ResponseEntity.ok(null);
+        }
+        System.out.println("Item does NOT belong to a user: ");
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @GetMapping("/{id}")  // localhost:8080/item/1
-    public ListItem getItemById(@PathVariable int id) {
-        return listItemService.getItemById(id);
+    public ResponseEntity<ListItem> getItemById(@PathVariable int id, Authentication authentication) {
+        // Get the token
+        JwtAuthenticationToken jwtAuthenticationToken = (JwtAuthenticationToken) authentication;
+        String jwtToken = Functions.extractJwtToken(jwtAuthenticationToken);
+        // Get the userId
+        int userId = TokenService.getUserIdFromToken(jwtToken);
+        // Authorize
+        if (hasUserAccessToList(userId, id)) {
+            System.out.println("Item does belong to a user: ");
+            com.example.Java4.entity.ListItem item = listItemService.getItemById(id);
+            return ResponseEntity.ok(item);
+        }
+        System.out.println("Item does NOT belong to a user: ");
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
-    @PutMapping("/")  // localhost:8080/item
-    public ListItem updateListItem(@RequestBody ListItem item) {
-        return listItemService.updateListItem(item);
+    @PutMapping("/")  // localhost:8080/item/
+    public ResponseEntity<ListItem> updateListItem(@RequestBody ListItem item, Authentication authentication) {
+        // Get the token
+        JwtAuthenticationToken jwtAuthenticationToken = (JwtAuthenticationToken) authentication;
+        String jwtToken = Functions.extractJwtToken(jwtAuthenticationToken);
+        // Get the userId
+        int userId = TokenService.getUserIdFromToken(jwtToken);
+        int id = item.getId();
+        // Authorize
+        if (hasUserAccessToList(userId, id)) {
+            System.out.println("Item does belong to a user: ");
+            listItemService.updateListItem(item);
+            return ResponseEntity.ok(item);
+        }
+        System.out.println("Item does NOT belong to a user: ");
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
+
+    private boolean hasUserAccessToList(int userId, int listId) {
+        // Check if the user has access to the list
+        return listService.doesListBelongToUser(listId , userId);
+    }
+    private boolean doesItemBelongToList(int itemId, int listId) {
+        // Check if the item belongs to the list
+        return listItemService.doesItemBelongToList(itemId, listId);
     }
 }
